Update infrastructure namespaces to providers
