+++ Guppi.Infrastructure/Providers/Git/GitProvider.cs
@@ -1,0 +1,30 @@
+using Guppi.Core.Exceptions;
+using Guppi.Core.Interfaces;
+using LibGit2Sharp;
+
+namespace Guppi.Core.Providers.Git;
+
+public class GitProvider(IProcessProvider process) : IGitProvider
+{
+    const string WorkingDirectory = ".";
+    private readonly IProcessProvider _process = process;
+
+    public void RunGit(string args)
+    {
+        _process.Start("git", args, WorkingDirectory, true);
+    }
+
+    public void SwitchToBranch(string branchName)
+    {
+        try
+        {
+            using var repo = new Repository(WorkingDirectory);
+            var branch = repo.Branches[branchName] ?? throw new WarningException($"The branch {branchName} does not exist.");
+            Commands.Checkout(repo, branch);
+        }
+        catch (RepositoryNotFoundException)
+        {
+            throw new UnconfiguredException("The current directory is not a git repository.");
+        }
+    }
+}
